llm-performance-viz/pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "llm_test_tool"
version = "0.1.0"
description = "A lightweight tool for testing LLM API performance"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
dependencies = [
    "requests>=2.25.0",
    "PyYAML>=6.0",
    "fastapi>=0.68.0",
    "uvicorn[standard]>=0.15.0",
    "pandas>=1.3.0",
    "numpy>=1.21.0",
    "python-multipart>=0.0.5",
    "matplotlib>=3.5.0",
    "seaborn>=0.11.0",
]

[project.scripts]
llm-test = "llm_test_tool.main:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.black]
line-length = 88
target-version = ["py38"]

[tool.isort]
profile = "black"
line_length = 88
